separate completeness file
